+++ doomsday/engine/portable/include/rend_list.h
@@ -53,7 +53,7 @@ void            RL_DeleteLists();
 void            RL_AddPoly(rendpoly_t *poly);
 void            RL_PreparePlane(subplaneinfo_t *plane, rendpoly_t *poly,
                                 float height, subsector_t *subsector);
-void            RL_VertexColors(rendpoly_t *poly, int lightlevel,
+void            RL_VertexColors(rendpoly_t *poly, float lightlevel,
                                 const byte *rgb);
 void            RL_RenderAllLists();
